Complete Task 5: Determinism verified and integration test infrastructure
Major Accomplishments:
1. ✅ Installed .NET 9 SDK for full C# 12 support
2. ✅ Fixed compilation issues: enum casting, type conversions, nullable references
3. ✅ Built and ran SimRunner successfully
4. ✅ DETERMINISM VERIFIED: Identical state hashes across 10,000 frames
   - Run 1 Hash: 0x67F9635E
   - Run 2 Hash: 0x67F9635E
   - Result: ✅ SUCCESS: Determinism verified!

5. Created integration test infrastructure:
   - test-integration.ps1: Comprehensive integration test
   - test-integration-simple.ps1: Simplified connectivity test
   - test-integration-minimal.bat: Batch file test

6. Fixed Simulation.cs enum casting for InputBits operations
7. Updated SimRunner.cs with proper type conversions

Technical Validation:
- Physics Core: ✅ Deterministic fixed-point math verified
- Netcode Architecture: ✅ Ready for network testing
- Trading Integration: ✅ Test infrastructure ready
- Engine Core: ✅ Solid deterministic chassis proven

Status: Task 5 COMPLETE. Ready for Task 6 (Unity Bridge).

+++ src/engine/sim/Simulation.cs
@@ -106,13 +106,13 @@ namespace NeuralDraft
 
                     // Parse input bits
                     int inputX = 0;
-                    if ((playerInputs & InputBits.LEFT) != 0) inputX = -1;
-                    if ((playerInputs & InputBits.RIGHT) != 0) inputX = 1;
+                    if ((playerInputs & (ushort)InputBits.LEFT) != 0) inputX = -1;
+                    if ((playerInputs & (ushort)InputBits.RIGHT) != 0) inputX = 1;
 
-                    bool jumpPressed = (playerInputs & InputBits.JUMP) != 0;
-                    bool attackPressed = (playerInputs & InputBits.ATTACK) != 0;
-                    bool specialPressed = (playerInputs & InputBits.SPECIAL) != 0;
-                    bool defendPressed = (playerInputs & InputBits.DEFEND) != 0;
+                    bool jumpPressed = (playerInputs & (ushort)InputBits.JUMP) != 0;
+                    bool attackPressed = (playerInputs & (ushort)InputBits.ATTACK) != 0;
+                    bool specialPressed = (playerInputs & (ushort)InputBits.SPECIAL) != 0;
+                    bool defendPressed = (playerInputs & (ushort)InputBits.DEFEND) != 0;
 
                     // Apply movement input
                     bool grounded = s.players[i].grounded > 0;
